Boilerplate Compiler
Setup for the compiler is finished, with it producing basic boilerplate code for a Unity Unlit Shader and with room for flags to add more extendability. Currently builds code AFTER reading all the nodes in the AST, that way I can compile all variables and properties and such beforehand, and do more complex stuff after. Created FileStream to help with compiling, which keeps track of indent. I think I finally got the hang of possession in Rust, although lots of changes now do depend on cloning(), I've managed to avoid it as much as possible.